requirements.txt
keras==2.8.0
numpy==1.21.4
streamlit==1.18.0
opencv-python-headless==4.5.3
tensorflow==2.7.0
Pillow==9.0.0
